Broke up __init__ into multiple files
